ACTUALIZACION

+++ Models/Ventas.cs
@@ -16,7 +16,7 @@ namespace Raldy_P2_APL2.Models
         public DateTime Fecha { get; set; }
         public int ClienteId { get; set; }
         public float Monto { get; set; }
-        public float Balance { get; set; }
+        public double Balance { get; set; }
         public bool CobroNoPerdiente { get; set; }
         public double Cobrado { get; set; }
         public virtual Clientes Clientes { get; set; }
